BUG: replace with regex raising for StringDType (#41343)
